refactor: Update avatar URL and user ID for existing users in GitHub and Google strategies
Update the GitHub and Google authentication strategies to handle existing users. This includes updating the avatar URL for users who don't have an avatar and updating the user ID for users who don't have a GitHub or Google ID. These changes improve the consistency and accuracy of user data in the application.