refactor: Add agreeContract route and controller function
Add the agreeContract route and controller function to handle the agreement of the contract by a user. This route allows a user to agree to the contract by updating the isAgreed field in the user model to true. By adding this functionality, we ensure that users can indicate their agreement to the contract, improving compliance and legal requirements.